Fix sandbox import cycle
The sandbox internals depended on public sandbox models, which
pulled package initialization back through the public API and left
`vercel._internal.sandbox` partially initialized.

Move sandbox model definitions into `_internal` and re-export them
from the public package so the public API stays stable while
internal imports no longer route through `vercel.sandbox`.
